requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
closure==20180204
colorama==0.4.1
css==0.1
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-angular2==0.0.1
django-angularjs==1.0.2
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-embed-video==1.2.0
django-phone-field==1.7.0
django-sb-admin==0.3.7
django-static==1.5.6
django-static-templates==0.0.1
django-staticfiles==1.2.1
django-videofield==0.1.1
djangocms-glyphicon-awesome==0.1.0
djangorestframework==3.10.3
entrypoints==0.3
fanstatic==1.1
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jquery.javascript==1.0.0
js.angular==1.1.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==5.4.1
post==2019.3.22
prometheus-client==0.6.0
prompt-toolkit==2.0.9
Pygments==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-gettext==4.0
pytz==2018.9
pywinpty==0.5.5
pyzmq==18.0.1
q==2.6
qtconsole==4.4.3
requests==2.21.0
Send2Trash==1.5.0
shutilwhich==1.1.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
transaction==2.4.0
urllib3==1.24.1
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
widgetsnbextension==3.4.2
zope.browser==2.3
zope.browserpage==4.3.0
zope.component==4.5
zope.configuration==4.3.1
zope.contentprovider==4.2.1
zope.contenttype==4.4
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.event==4.4
zope.exceptions==4.3
zope.hookable==4.2.0
zope.i18n==4.6.2
zope.i18nmessageid==4.3.1
zope.interface==4.6.0
zope.location==4.2
zope.pagetemplate==4.4.1
zope.proxy==4.3.1
zope.publisher==5.0.1
zope.schema==4.9.3
zope.security==4.3.1
zope.tal==4.4
zope.tales==5.0
zope.traversing==4.3.1
zope.viewlet==4.2.1
